Move ids and use them in more places
