+++ bindings/python/iota_sdk/__init__.py
@@ -32,7 +32,6 @@ from .types.feature import *
 from .types.irc_27 import *
 from .types.irc_30 import *
 from .types.filter_options import *
-from .types.ids import *
 from .types.input import *
 from .types.native_token import *
 from .types.network_info import *
