+++ bindings/python/iota_sdk/types/block/block.py
@@ -6,13 +6,17 @@ from dataclasses import dataclass, field
 from typing import Any, Dict, TypeAlias, Union
 from dataclasses_json import config
 from iota_sdk.utils import Utils
-from iota_sdk.types.common import HexStr, json, SlotIndex
-from iota_sdk.types.ids import SlotCommitmentId
+from iota_sdk.types.common import HexStr, IdWithSlotIndex, json, SlotIndex
 from iota_sdk.types.node_info import ProtocolParameters
 from iota_sdk.types.signature import Signature
 from iota_sdk.types.block.body.basic import BasicBlockBody
 from iota_sdk.types.block.body.validation import ValidationBlockBody
 from iota_sdk.types.block.body.type import BlockBodyType
+
+
+class BlockId(IdWithSlotIndex):
+    """Represents a block ID.
+    """
 
 
 @json
@@ -86,7 +90,7 @@ class Block:
     ))
     signature: Signature
 
-    def id(self, params: ProtocolParameters) -> HexStr:
+    def id(self, params: ProtocolParameters) -> BlockId:
         """Returns the block ID as a hexadecimal string.
         """
         return Utils.block_id(self, params)
