+++ bindings/python/iota_sdk/types/block/body/basic.py
@@ -5,8 +5,9 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import List, Optional
 from dataclasses_json import config
+from iota_sdk.types.block.block import BlockId
 from iota_sdk.types.block.body.type import BlockBodyType
-from iota_sdk.types.common import HexStr, json
+from iota_sdk.types.common import json
 from iota_sdk.types.payload import Payload, deserialize_payload
 
 
@@ -26,12 +27,12 @@ class BasicBlockBody:
     type: int = field(
         default_factory=lambda: int(BlockBodyType.Basic),
         init=False)
-    strong_parents: List[HexStr]
+    strong_parents: List[BlockId]
     max_burned_mana: int = field(metadata=config(
         encoder=str
     ))
-    weak_parents: Optional[List[HexStr]] = None
-    shallow_like_parents: Optional[List[HexStr]] = None
+    weak_parents: Optional[List[BlockId]] = None
+    shallow_like_parents: Optional[List[BlockId]] = None
     payload: Optional[Payload] = field(default=None, metadata=config(
         decoder=deserialize_payload
     ))
