--- bindings/python/iota_sdk/types/ids.py
@@ -1,31 +1,0 @@
-# Copyright 2024 IOTA Stiftung
-# SPDX-License-Identifier: Apache-2.0
-
-class IdWithSlotIndex(str):
-    """Represents an hex encoded ID that contains a slot index at the end.
-
-    Attributes:
-        id: The hex encoded ID with a slot index.
-
-    """
-
-    def slot_index(self):
-        """Returns the slot index of the ID.
-        """
-        return int.from_bytes(
-            bytes.fromhex(self[-8:]), 'little')
-
-
-class BlockId(IdWithSlotIndex):
-    """Represents a block ID.
-    """
-
-
-class TransactionId(IdWithSlotIndex):
-    """Represents a transaction ID.
-    """
-
-
-class SlotCommitmentId(IdWithSlotIndex):
-    """Represents a slot commitment ID.
-    """
