+++ bindings/python/iota_sdk/types/output_data.py
@@ -5,8 +5,9 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Optional
 from iota_sdk.types.address import Address
-from iota_sdk.types.common import HexStr, json
+from iota_sdk.types.common import json
 from iota_sdk.types.output import Output
+from iota_sdk.types.output_id import OutputId
 from iota_sdk.types.output_metadata import OutputMetadata
 from iota_sdk.types.signature import Bip44
 
@@ -25,7 +26,7 @@ class OutputData():
         remainder: Whether the output represents a remainder amount.
         chain: A list of chain state indexes.
     """
-    output_id: HexStr
+    output_id: OutputId
     metadata: OutputMetadata
     output: Output
     address: Address
