+++ bindings/python/iota_sdk/types/output_id.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from iota_sdk.types.common import HexStr
-from iota_sdk.types.ids import TransactionId
+from iota_sdk.types.transaction_data import TransactionId
 
 
 class OutputId(dict):
